define Grant pydantic model, refactor frontend to use it
