+++ frontend/src/app/dataReducer.ts
@@ -3,7 +3,7 @@ import { Aggregate } from 'api/models/Aggregate';
 import { DivisionAggregate } from 'api/models/DivisionAggregate';
 import { loadAbstract, loadData, loadGrants, loadRelated, loadTypeahead } from './actions';
 import { addChips, deleteChip, setGrantOrder, setTerms } from './filterReducer';
-import { Grant } from './types';
+import { Grant } from '../api/models/Grant';
 
 type GrantState = {
   perYear: Aggregate[],
